Getting the edit product to work

+++ deleteproduct.php
@@ -1,17 +1,18 @@
+<script type="application/javascript">
+if(confirm('Are you sure you want to delete this record')){
+    <?php $continue = TRUE; ?>
+}
+</script>
 
 <?php
-$db = new PDO('mysql:host=us-cdbr-azure-southcentral-e.cloudapp.net;dbname=inventoryms;charset=utf8mb4', 'bee886bc8793e7', '362289e3',array(PDO::ATTR_EMULATE_PREPARES => false,
-    PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION));
+if($continue){
+    $db = new PDO('mysql:host=us-cdbr-azure-southcentral-e.cloudapp.net;dbname=inventoryms;charset=utf8mb4', 'bee886bc8793e7', '362289e3',array(PDO::ATTR_EMULATE_PREPARES => false,
+        PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION));
 
     if(isset($_GET['prodid'])){
-    $prodId = $_GET['prodid'];
+        $prodId = $_GET['prodid'];
 
         try {
-            echo "<script>";
-            echo "if(!confirm('Are you sure you want to delete this record')){";
-            echo "window.location = 'allproducts.php'";
-            echo "}";
-            echo "</script>";
             $sql = "DELETE FROM addproduct
                     WHERE productid = {$prodId}";
             $sth = $db->query($sql);
@@ -24,4 +25,8 @@ $db = new PDO('mysql:host=us-cdbr-azure-southcentral-e.cloudapp.net;dbname=inven
     }else{
         header("Location: allproducts.php");
     }
+}else{
+    header("Location: allproducts.php");
+}
+
 ?>
